Update ThanksDuka.lua
Modified EndRoll to properly handle duplicate items. ThanksDuka now awards one item to the highest roller, the next duplicate to the next highest, and so on.

Addon still improperly leaves the duplicate items on the loot window and user must press end roll for duplicate item. It might also skip a second player in the event of a tie.