Update ThanksDuka.lua
Fixed a UI bug where the hyperlink (tooltip) of items on the loot tab was above the text. Now users can properly hover over the item's name to see the tooltip.

Added a function in the loot history tab to deal with manually awarding an item to another person in the raid in event of accidental rolls or changes of mind.

Added debugging functions (off by default) for testing the addon without being in a raid group.